% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbs.R
\name{attempt}
\alias{attempt}
\title{Attempt}
\usage{
attempt(expr, msg = NULL, verbose = FALSE, silent = FALSE)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{msg}{the message to return if an error or warning occurs}

\item{verbose}{whether to return the expression producing the error}

\item{silent}{whether the error should be silent}
}
\description{
A wrapper around base try that allows you to set a custom message when an error/warning occurs.
\code{attempt} returns the value if there are no errors or messages.
}
\examples{
\dontrun{
attempt(log("a"), msg = "Nop !")
}
}
