% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbs.R
\name{discretly}
\alias{discretly}
\title{discretly}
\usage{
discretly(.f)
}
\arguments{
\item{.f}{the function to wrap}
}
\value{
an error if any, a warning if any, the result if any
}
\description{
Prevent a funtion from printing message or warning
}
\examples{
\dontrun{
discrete_mat <- discretly(matrix)
discrete_mat(1:3, 2)
}
}
