% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if.R
\name{if_else}
\alias{if_else}
\title{If this, then that, else that}
\usage{
if_else(.x, .p = isTRUE, .f, .else)
}
\arguments{
\item{.x}{the object to test. If \code{NULL} (the default), only .p is evaluated.}

\item{.p}{the predicate for testing. Defaut is \code{isTRUE}.}

\item{.f}{what to do if TRUE}

\item{.else}{what to do if TRUE}
}
\value{
the evalution
}
\description{
If this, then that, else that
}
\examples{
\dontrun{
a <- if_else(.x = curl::has_internet(), .f = ~ "Yay", .else = ~ "Nay")
}

}
