% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_class.R
\name{is_try_error}
\alias{is_try_error}
\title{Is the element of class "try-error"?}
\usage{
is_try_error(.x)
}
\arguments{
\item{.x}{the object to test}
}
\value{
A logical
}
\description{
Is the element of class "try-error"?
}
\examples{
x <- attempt(log("a"), silent = TRUE)
is_try_error(x)

}
