% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbs.R
\name{with_message}
\alias{with_message}
\alias{with_warning}
\alias{without_message}
\alias{without_warning}
\title{Manipulate messages and warnings}
\usage{
with_message(.f, msg)

with_warning(.f, msg)

without_message(.f)

without_warning(.f)
}
\arguments{
\item{.f}{the function to wrap}

\item{msg}{the message to print}
}
\value{
a function
}
\description{
\code{with_message} and \code{with_warning} add a warning or a message to a function.
\code{without_message} and \code{without_warning} turn the warning and message off.
}
\examples{
msg_as_num <- with_message(as.numeric, msg = "Numeric conversion")
warn_as_num <- with_warning(as.numeric, msg = "Numeric conversion")
}
