% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbs.R
\name{silent_attempt}
\alias{silent_attempt}
\title{Silently attempt}
\usage{
silent_attempt(...)
}
\arguments{
\item{...}{the expression to evaluate}
}
\value{
an error if any, a warning if any.
}
\description{
A wrapper around silently and attempt
}
\examples{
\dontrun{
silent_attempt(warn("nop!"))
}
}
