% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch.R
\name{silently}
\alias{silently}
\title{Silently}
\usage{
silently(expr)
}
\arguments{
\item{expr}{the expression to be evaluated}
}
\value{
an error if any, a warning if any.
}
\description{
silently returns an error or a warning if any,
else returns nothing.
}
\examples{
\dontrun{
silently(warn("nop!"))
}
}
