% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch.R
\name{surely}
\alias{surely}
\title{surely}
\usage{
surely(.f)
}
\arguments{
\item{.f}{the function to wrap}
}
\value{
an error if any, a warning if any, the result if any
}
\description{
Wrap a function in a try
}
\examples{
\dontrun{
sure_log <- surely(log)
sure_log(1)
sure_log("a")
}
}
