% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch.R
\name{try_catch}
\alias{try_catch}
\title{Try Catch}
\usage{
try_catch(expr, .e = NULL, .w = NULL, .f = NULL)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{.e}{a one side formula or a function evaluated when an error occurs}

\item{.w}{a one side formula or a function evaluated when a warning occurs}

\item{.f}{a one side formula or an expression evaluated before returning or exiting}
}
\description{
A friendlier try catch function
}
\examples{

try_catch(log("a"), .e = ~ paste0("There was an error: ", .x))
}
