% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch.R
\name{try_that}
\alias{try_that}
\title{Try that}
\usage{
try_that(expr, msg = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{msg}{the message to return if an error occurs}

\item{verbose}{wether or not to return to expression producing the error}
}
\description{
A wrapper around base try that allows you to set a custom message when an error occurs
}
\examples{

try_that(log("a"), msg = "Nop !")

}
