% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempt-package.R
\docType{package}
\name{attempt-package}
\alias{attempt-package}
\title{attempt package}
\description{
A friendlier condition handler, inspired by \code{purrr} mappers and based on \code{rlang}.
'attempt' extends and facilitates condition handling by providing a consistent grammar, and
provides a set of easy to use functions for common tests and conditions. 'attempt' only depends
on \code{rlang}, and focuses on speed, so it can be easily integrated in other functions and used
in data analysis.
}
\author{
colin <contact@colinfay.me>
}
