% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempt-package.R
\docType{package}
\name{attempt-package}
\alias{attempt-package}
\title{attempt package}
\description{
Tools for defensive programming in R.
'attempt' extends and facilitates defensive programming by providing a consistent grammar, and
provides a set of easy to use functions for common tests and conditions. 'attempt' only depends
on \code{rlang} & \code{withr}, and focuses on speed, so it can be easily integrated in other functions and used
in data analysis.
}
\author{
colin <contact@colinfay.me>
}
