% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempt-package.R
\docType{package}
\name{attempt-package}
\alias{attempt-package}
\title{attempt package}
\description{
Friendlier Condition Handler for R, inspired by \code{purrr} mappers and based on \code{rlang}.
}
\author{
colin <contact@colinfay.me>
}
