% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch.R
\name{attempt}
\alias{attempt}
\title{Attempt}
\usage{
attempt(expr, msg = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{msg}{the message to return if an error occurs}

\item{verbose}{wether or not to return to expression producing the error}
}
\description{
Wrappers around base try that allows you to set a custom message when an error/warning occurs.
\code{attempt} returns the value if there is no error nor message. \code{silent_attempt} stays
silent unless error or warning.
}
\examples{
\dontrun{
attempt(log("a"), msg = "Nop !")
}
}
