% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if.R
\name{if_else}
\alias{if_else}
\title{If this, then that, else that}
\usage{
if_else(.x = NULL, .p = isTRUE, .f, .else)
}
\arguments{
\item{.x}{the object to test. If \code{NULL} (the default), only .p is evaluated.}

\item{.p}{the predicate for testing. Defaut is \code{isTRUE}.}

\item{.f}{what to do if TRUE}

\item{.else}{what to do if TRUE}
}
\value{
the evalution
}
\description{
If this, then that, else that
}
\note{
If you need to call a function that takes no argument at \code{.p} (like \code{curl::has_internet()}),
you can set \code{.x} as \code{NULL}, which is also the default. If ever you don't pass anything to \code{.x}, don't
forget to name the arguments.
}
\examples{
\dontrun{
a <- if_else(.f = curl::has_internet, ~ "Yay", ~ "Nay")
}

}
