% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if.R
\name{if_else}
\alias{if_else}
\title{If this, then do that}
\usage{
if_else(.x, .p = isTRUE, .f, .else)
}
\arguments{
\item{.x}{the object to test}

\item{.p}{the predicate for testing. Defaut is \code{isTRUE}}

\item{.f}{what to do if TRUE}

\item{.else}{what to do if TRUE}
}
\value{
the evalution
}
\description{
If this, then do that
}
\examples{
a <- if_else(1, is.numeric, ~ return("Yay"), ~ return("Nay"))
}
