% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if.R
\name{if_then}
\alias{if_then}
\title{If this, then that}
\usage{
if_then(.x, .p = isTRUE, .f)
}
\arguments{
\item{.x}{the object to test}

\item{.p}{the predicate for testing. Defaut is \code{isTRUE}}

\item{.f}{what to do if TRUE}
}
\value{
the result in .f
}
\description{
If this, then that
}
\examples{
a <- if_then(1, is.numeric, ~ return("Yay"))
a <- if_then(., ~ return(TRUE), ~ return("Yay"))
}
