% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbs.R
\name{with_message}
\alias{with_message}
\alias{with_warning}
\title{Add a message or warning to a function}
\usage{
with_message(.f, msg)

with_warning(.f, msg)
}
\arguments{
\item{.f}{the function to wrap}

\item{msg}{the message to print}
}
\value{
a function
}
\description{
Add a message or warning to a function
}
\examples{
msg_as_num <- with_message(as.numeric, msg = "Numeric conversion")
warn_as_num <- with_warning(as.numeric, msg = "Numeric conversion")
}
