% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch.R
\name{silently}
\alias{silently}
\title{Silently}
\usage{
silently(.f)
}
\arguments{
\item{.f}{the function to silence}
}
\value{
an error if any, a warning if any. The result is never returned.
}
\description{
silently returns a new function that will returns an error or a warning if any,
or else returns nothing.
}
\examples{
\dontrun{
silent_log <- silently(log)
silent_log(1)
silent_log("a")
}
}
